requirements.txt
flask
flask-limiter
flask-talisman
resend
python-dotenv
gunicorn
